TestData/Core/basic_cases.dat

#name
ParseConstSection.Basic
#data
unit Unit1;
interface
const
   A = 1;
	B= 2;
   C =3;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="Unit1"
| 		ntQualifiedIdentifier anName="Unit1"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="A" anValueText="1"
| 					ntIdentifier anName="A"
| 					ntExpression anValueText="1"
| 				ntConstant anName="B" anValueText="2"
| 					ntIdentifier anName="B"
| 					ntExpression anValueText="2"
| 				ntConstant anName="C" anValueText="3"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="3"
|		ntImplementation



#name
ParseResStringSection.Basic
#data
unit TestCase.ParseStringSection;
interface
resourcestring
	SProduct = 'Hello world!';
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="TestCase.ParseStringSection"
| 		ntQualifiedIdentifier anName="TestCase.ParseStringSection"
| 		ntInterfaceSection
| 			ntResourceStrings
| 				ntResourceString anName="SProduct" anValueText="Hello world!"
| 					ntExpression anValueText="Hello world!"
| 		ntImplementation



#name
ParseResStringSection.Concatenated
#data
unit TestCase.ParseStringSection;
interface
resourcestring
	SProduct = 'Hello, '+'world!';
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="TestCase.ParseStringSection"
| 		ntQualifiedIdentifier anName="TestCase.ParseStringSection"
| 		ntInterfaceSection
| 			ntResourceStrings
| 				ntResourceString anName="SProduct" anValueText="Hello, world!"
| 					ntExpression anValueText="Hello, world!"
| 		ntImplementation




#name
MethodInventory.ParseAndIsPossible
#data
unit Inventory;
interface

implementation
end.
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="Inventory"
| 		ntQualifiedIdentifier
| 		ntInterfaceSection
| 		ntImplementation

#name
ParseProperty
#data
unit U;
interface
type
  T = class
  private
    F: Integer;
  public
    property P: Integer read F write F;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntPrivate
| 							ntField
| 								ntName anName="F"
| 								ntType anName="Integer"
| 						ntPublic
| 							ntProperty
| 								ntProperty
| 									ntIdentifier anName="P"
| 									ntType anName="Integer"
| 									ntPropertySpecifiers
| 										ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 											ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 												ntQualifiedIdentifier
| 												ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntImplementation

#name
ParsePropertyDirective
#data
unit U;
interface
type
  T = class
  private
    F: Integer;
  public
    property P: Integer read F write F;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntPrivate
| 							ntField
| 								ntName anName="F"
| 								ntType anName="Integer"
| 						ntPublic
| 							ntProperty
| 								ntProperty
| 									ntIdentifier anName="P"
| 									ntType anName="Integer"
| 									ntPropertySpecifiers
| 										ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 											ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 												ntQualifiedIdentifier
| 												ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntImplementation

#name
ParseStream
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
ParseMethodOrProperty
#data
unit U;
interface
type
  T = class
    procedure P;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntMethod
| 							ntMethod anKind="avProcedure"
| 								ntQualifiedIdentifier anName="P"
| 		ntImplementation

#name
ParseFieldDecl
#data
unit U;
interface
type
  T = class
    A, B: Integer;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntField
| 							ntName anName="A"
| 							ntName anName="B"
| 							ntType anName="Integer"
| 		ntImplementation

#name
ParseCore
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
ParseFile
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
ParseText
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
ParseLibraryFile
#data
library L;
begin end.
#errors
#document
| ntCompilationUnit
| 	ntLibrary
| 		ntIdentifier anName="L"
| 		ntUnknown anName="Not implemented yet" anMissing="avTrue"

#name
ParsePackageFile
#data
package P;
end.
#errors
#document
| ntCompilationUnit
| 	ntPackage
| 		ntQualifiedIdentifier

#name
ParseProgramFile
#data
program P;
begin end.
#errors
#document
| ntCompilationUnit
| 	ntProgram
| 		ntIdentifier anName="P"
| 		ntUnknown anName="Not implemented yet" anMissing="avTrue"

#name
ParseUnitDeclaration
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
ParseUsesClause
#data
unit U;
interface
uses System.SysUtils;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntUses
| 				ntUsedUnit anName="System.SysUtils"
| 					ntQualifiedIdentifier anName="System.SysUtils"
| 		ntImplementation

#name
ParseUsedUnit
#data
unit U;
interface
uses System.SysUtils;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntUses
| 				ntUsedUnit anName="System.SysUtils"
| 					ntQualifiedIdentifier anName="System.SysUtils"
| 		ntImplementation

#name
ParseImplementationDecl
#data
unit U;
interface
implementation
const
  C = 1;
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation
| 			ntConstants
| 				ntConstant anName="C" anValueText="1"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1"

#name
ParseQualifiedIdentifier
#data
unit A.B;
interface
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="A.B"
| 		ntQualifiedIdentifier anName="A.B"
| 		ntInterfaceSection
| 		ntImplementation

#name
ParseTerm
#data
unit U;
interface
const
  C = 1 + 2 * 3;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1+2*3"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1+2*3"
| 		ntImplementation

#name
ParseSimpleExpression
#data
unit U;
interface
const
  C = 1 + 2 * 3;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1+2*3"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1+2*3"
| 		ntImplementation

#name
ParseExpression
#data
unit U;
interface
const
  C = 1 + 2 * 3;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1+2*3"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1+2*3"
| 		ntImplementation

#name
ParseFinalizationSection
#data
unit U;
interface
implementation
initialization
finalization
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation
| 		ntInitialization
| 			ntStatements
| 				ntUnknown anName="Not implemented yet" anMissing="avTrue"

#name
ParseEnumeratedType
#data
unit U;
interface
type
  TEnum = (A, B);
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TEnum"
| 					ntType anName="avEnum"
| 						ntIdentifier anName="A"
| 						ntIdentifier anName="B"
| 		ntImplementation

#name
ParseSubrangeType
#data
unit U;
interface
type
  TSub = 1..2;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TSub"
| 					ntType anName="1"
| 						ntType
| 							ntExpression
| 								ntExpression
| 									ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 										ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntImplementation

#name
ParseRealType
#data
unit U;
interface
type
  TReal = 1.5;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TReal"
| 					ntType anName="1.5"
| 						ntType
| 		ntImplementation

#name
ParseOrdinalType
#data
unit U;
interface
type
  TOrd = 1;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TOrd"
| 					ntType anName="1"
| 						ntType
| 							ntExpression
| 								ntExpression
| 									ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 										ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntImplementation

#name
ParseVariableReference
#data
unit U;
interface
implementation
procedure P;
begin
  A.B;
end;
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation
| 			ntMethod
| 				ntName anName="procedure"
| 					ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 						ntUnknown anName="E2123 PROCEDURE, FUNCTION, PROPERTY, or VAR expected" anMissing="true"
| 					ntName
| 					ntStatements
| 						ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 							ntEmptyStatement

#name
ParseVisibilitySection
#data
unit U;
interface
type
  T = class
  private
    F: Integer;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntPrivate
| 							ntField
| 								ntName anName="F"
| 								ntType anName="Integer"
| 		ntImplementation

#name
ParseVisibilitySectionContent
#data
unit U;
interface
type
  T = class
    F: Integer;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntField
| 							ntName anName="F"
| 							ntType anName="Integer"
| 		ntImplementation

#name
ParseInterfaceType
#data
unit U;
interface
type
  ITest = interface
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="ITest"
| 					ntUnknown anName="E2029 Type expected but interface found" anMissing="true"
| 		ntImplementation
| 	ntUnknown anName="Unexpected trailing tokens"

#name
ParseClassType
#data
unit U;
interface
type
  T = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 		ntImplementation

#name
ParseClassHelper
#data
unit U;
interface
type
  TStringHelper = class helper for string
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TStringHelper"
| 					ntType anType="avClass"
| 						ntUnknown anName="Skipping token to recover inside class body: helper" anMissing="true"
| 						ntUnknown anName="Skipping token to recover inside class body: for" anMissing="true"
| 						ntUnknown anName="Skipping token to recover inside class body: string" anMissing="true"
| 		ntImplementation

#name
ParseClassMemberList
#data
unit U;
interface
type
  T = record
    F: Integer;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="record"
| 						ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 						ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 						ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntImplementation
| 	ntUnknown anName="Unexpected trailing tokens"

#name
ParseProceduralType
#data
unit U;
interface
type
  TProc = procedure;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TProc"
| 					ntType anName="procedure"
| 		ntImplementation

#name
ParsePointerType
#data
unit U;
interface
type
  PInt = ^Integer;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="PInt"
| 					ntUnknown anName="E2029 Type expected but ^ found" anMissing="true"
| 		ntImplementation
| 	ntUnknown anName="Unexpected trailing tokens"

#name
ParseStringType
#data
unit U;
interface
type
  S10 = string[10];
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="S10"
| 					ntType
| 			ntUnknown anName="InvalidInterfaceDeclaration" anMissing="true"
| 		ntImplementation
| 	ntUnknown anName="Unexpected trailing tokens"

#name
ParseStructuredType
#data
unit U;
interface
type
  TInts = array of Integer;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TInts"
| 					ntType anType="array"
| 						ntType anType="array"
| 							ntType anName="Integer"
| 		ntImplementation

#name
ParseSimpleType
#data
unit U;
interface
type
  TAlias = 1;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TAlias"
| 					ntType anName="1"
| 						ntType
| 							ntExpression
| 								ntExpression
| 									ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 										ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntImplementation

#name
ParseTypeDeclaration
#data
unit U;
interface
type
  T = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 		ntImplementation

#name
ParseType
#data
unit U;
interface
type
  T = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 		ntImplementation

#name
ParseTypedConstant
#data
unit U;
interface
const
  C: array[0..1] of Integer = (1,2);
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="implementationend."
| 					ntIdentifier anName="C"
| 					ntType anType="array"
| 						ntType anType="array"
| 							ntExpression
| 								ntExpression
| 									ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 										ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 							ntUnknown anName="E2029 Type expected but ; found" anMissing="true"
| 					ntExpression anValueText="implementationend."
| 		ntImplementation

#name
ParseTypeId
#data
unit U;
interface
type
  TAlias = Integer;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TAlias"
| 					ntType anName="Integer"
| 		ntImplementation

#name
ParseConstantName
#data
unit U;
interface
const
  C = 1;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1"
| 		ntImplementation

#name
ParseMethodHeading
#data
unit U;
interface
type
  T = class
    procedure P;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntMethod
| 							ntMethod anKind="avProcedure"
| 								ntQualifiedIdentifier anName="P"
| 		ntImplementation

#name
ParseVarSection
#data
unit U;
interface
var
  V: Integer;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntVariables
| 				ntVariables
| 					ntIdentifier
| 						ntIdentifier
| 					ntUnknown anName="E2029 Type expected but var found" anMissing="true"
| 					ntUnknown anName="Expected PortabilityDirective but was V" anMissing="true"
| 					ntUnknown anName="Expected PortabilityDirective but was :" anMissing="true"
| 				ntVariables
| 					ntIdentifier
| 						ntIdentifier anName="Integer"
| 					ntUnknown anName="E2029 Type expected but ; found" anMissing="true"
| 					ntUnknown anName="Expected PortabilityDirective but was implementation" anMissing="true"
| 					ntUnknown anName="Expected PortabilityDirective but was end" anMissing="true"
| 		ntImplementation

#name
IsPossibleConstSection
#data
unit U;
interface
const
  C = 1;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1"
| 		ntImplementation

#name
IsPossibleMethodHeading
#data
unit U;
interface
type
  T = class
    procedure P;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntMethod
| 							ntMethod anKind="avProcedure"
| 								ntQualifiedIdentifier anName="P"
| 		ntImplementation

#name
IsPossibleResStringSection
#data
unit U;
interface
resourcestring
  S = 'x';
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntResourceStrings
| 				ntResourceString anName="S" anValueText="x"
| 					ntExpression anValueText="x"
| 		ntImplementation

#name
IsPossibleTypeDeclaration
#data
unit U;
interface
type
  T = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 		ntImplementation

#name
IsPossibleTypeSection
#data
unit U;
interface
type
  T = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 		ntImplementation

#name
IsPossibleUsesClause
#data
unit U;
interface
uses System.SysUtils;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntUses
| 				ntUsedUnit anName="System.SysUtils"
| 					ntQualifiedIdentifier anName="System.SysUtils"
| 		ntImplementation

#name
IsPossibleVarSection
#data
unit U;
interface
var
  V: Integer;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntVariables
| 				ntVariables
| 					ntIdentifier
| 						ntIdentifier
| 					ntUnknown anName="E2029 Type expected but var found" anMissing="true"
| 					ntUnknown anName="Expected PortabilityDirective but was V" anMissing="true"
| 					ntUnknown anName="Expected PortabilityDirective but was :" anMissing="true"
| 				ntVariables
| 					ntIdentifier
| 						ntIdentifier anName="Integer"
| 					ntUnknown anName="E2029 Type expected but ; found" anMissing="true"
| 					ntUnknown anName="Expected PortabilityDirective but was implementation" anMissing="true"
| 					ntUnknown anName="Expected PortabilityDirective but was end" anMissing="true"
| 		ntImplementation

#name
IsPossibleVisibilitySectionContent
#data
unit U;
interface
type
  T = class
    F: Integer;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntField
| 							ntName anName="F"
| 							ntType anName="Integer"
| 		ntImplementation

#name
ParseTypeSection
#data
unit U;
interface
type
  T = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 		ntImplementation

#name
ParseTypeParam
#data
unit U;
interface
type
  TGen<T> = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TGen"
| 					ntTypeParams
| 						ntTypeParam
| 							ntType anName="T"
| 					ntType anType="avClass"
| 		ntImplementation

#name
ParseTypeParamConstraint
#data
unit U;
interface
type
  TGen<T: class> = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TGen"
| 					ntTypeParams
| 						ntTypeParam
| 							ntType anName="T"
| 							ntConstraints
| 								ntClassConstraint
| 					ntType anType="avClass"
| 		ntImplementation

#name
ParseConstraintTypeRef
#data
unit U;
interface
type
  TGen<T: IInterface> = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TGen"
| 					ntTypeParams
| 						ntTypeParam
| 							ntType anName="T"
| 							ntConstraints
| 								ntType
| 									ntQualifiedIdentifier anName="IInterface"
| 					ntType anType="avClass"
| 		ntImplementation

#name
ParseTypeParams
#data
unit U;
interface
type
  TGen<T> = class end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TGen"
| 					ntTypeParams
| 						ntTypeParam
| 							ntType anName="T"
| 					ntType anType="avClass"
| 		ntImplementation

#name
ParseConstSection
#data
unit U;
interface
const
  C = 1;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1"
| 		ntImplementation

#name
ParseConstantDecl
#data
unit U;
interface
const
  C = 1;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1"
| 		ntImplementation

#name
IsPossibleConstantDecl
#data
unit U;
interface
const
  C = 1;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1"
| 		ntImplementation

#name
ParseResStringSection
#data
unit U;
interface
resourcestring
  S = 'x';
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntResourceStrings
| 				ntResourceString anName="S" anValueText="x"
| 					ntExpression anValueText="x"
| 		ntImplementation

#name
IsPossibleResourceStringDecl
#data
unit U;
interface
resourcestring
  S = 'x';
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntResourceStrings
| 				ntResourceString anName="S" anValueText="x"
| 					ntExpression anValueText="x"
| 		ntImplementation

#name
ParseResourceStringDecl
#data
unit U;
interface
resourcestring
  S = 'x';
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntResourceStrings
| 				ntResourceString anName="S" anValueText="x"
| 					ntExpression anValueText="x"
| 		ntImplementation

#name
ParseInterfaceDeclaration
#data
unit U;
interface
const
  C = 1;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntConstants
| 				ntConstant anName="C" anValueText="1"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1"
| 		ntImplementation

#name
ParseInitializationSection
#data
unit U;
interface
implementation
initialization
finalization
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation
| 		ntInitialization
| 			ntStatements
| 				ntUnknown anName="Not implemented yet" anMissing="avTrue"

#name
ParseImplementationSection
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
ParseInterfaceSection
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
ParsePortabilityDirective
#data
unit U platform;
interface
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U" anPlatform="avTrue" anExperimental="avTrue"
| 		ntQualifiedIdentifier anName="U"
| 		ntPortabilityDirective anPlatform="avTrue"
| 		ntInterfaceSection
| 		ntImplementation
| 	ntUnknown anName="Unexpected trailing tokens"

#name
IsPossiblePortabilityDirective
#data
unit U platform;
interface
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U" anPlatform="avTrue" anExperimental="avTrue"
| 		ntQualifiedIdentifier anName="U"
| 		ntPortabilityDirective anPlatform="avTrue"
| 		ntInterfaceSection
| 		ntImplementation
| 	ntUnknown anName="Unexpected trailing tokens"

#name
IsPossibleProperty
#data
unit U;
interface
type
  T = class
  private
    F: Integer;
  public
    property P: Integer read F write F;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntPrivate
| 							ntField
| 								ntName anName="F"
| 								ntType anName="Integer"
| 						ntPublic
| 							ntProperty
| 								ntProperty
| 									ntIdentifier anName="P"
| 									ntType anName="Integer"
| 									ntPropertySpecifiers
| 										ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 											ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 												ntQualifiedIdentifier
| 												ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntImplementation

#name
ParseIdentifierList
#data
unit U;
interface
type
  T = class
    A, B: Integer;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntField
| 							ntName anName="A"
| 							ntName anName="B"
| 							ntType anName="Integer"
| 		ntImplementation

#name
ParseScriptFile
#data
A := 1;
#errors
#document
| ntCompilationUnit
| 	ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntUnknown anName="E2029 Top-level declaration expected but ; found" anMissing="true"

#name
ParseIndexSpecifier
#data
unit U;
interface
type
  T = class
  private
    F: Integer;
  public
    property P: Integer index 1 read F write F;
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="T"
| 					ntType anType="avClass"
| 						ntPrivate
| 							ntField
| 								ntName anName="F"
| 								ntType anName="Integer"
| 						ntPublic
| 							ntProperty
| 								ntProperty
| 									ntIdentifier anName="P"
| 									ntType anName="Integer"
| 									ntPropertySpecifiers
| 										ntExpression
| 											ntExpression
| 												ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 													ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 		ntImplementation

#name
ParseIdentifier
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
IsPossibleIdentifier
#data
unit U;
interface

implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation

#name
IsPossibleImplementationDecl
#data
unit U;
interface
implementation
const
  C = 1;
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation
| 			ntConstants
| 				ntConstant anName="C" anValueText="1"
| 					ntIdentifier anName="C"
| 					ntExpression anValueText="1"

#name
IsPossibleLabelID
#data
unit U;
interface
implementation
procedure P;
label L1;
begin
  goto L1;
L1:
end;
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 		ntImplementation
| 			ntMethod
| 				ntName anName="procedure"
| 					ntUnknown anName="Not implemented yet" anMissing="avTrue"
| 						ntUnknown anName="E2123 PROCEDURE, FUNCTION, PROPERTY, or VAR expected" anMissing="true"
| 					ntName
| 					ntUnknown anName="xxxxx Expected block" anMissing="true"
| 	ntUnknown anName="Unexpected trailing tokens"

#name
ParseAncestorList
#data
unit U;
interface
type
  TChild = class(TObject, IInterface)
  end;
implementation
end.
#errors
#document
| ntCompilationUnit
| 	ntUnitDeclaration anName="U"
| 		ntQualifiedIdentifier anName="U"
| 		ntInterfaceSection
| 			ntTypeSection
| 				ntTypeDecl anName="TChild"
| 					ntType anType="avClass"
| 						ntAncestorList
| 							ntQualifiedIdentifier anName="TObject"
| 							ntQualifiedIdentifier anName="IInterface"
| 		ntImplementation
